renderer: egl: allow GL textures to be exposed.
When the EGL_MESA_drm_image extension is used for video processing
from a VA surface to an EGL image, add the -m mesa_texture option
to allow the 2D texture to be created and managed by the EGL client
code.

So far, this only works with GL_RGBA textures.